divr_llm/requirements.txt
pyyaml>=6.0.1
torch==2.5.1
torchaudio==2.5.1
accelerate>=0.34.0
peft>=0.14.0
matplotlib>=3.10.0
pandas>=2.3.0
scikit-learn>=1.7.0
transformers @ git+https://github.com/huggingface/transformers